pass estimated outage of the backend to the DashboardSyncCardDisplayerImpl.kt
Former-commit-id: 793bfaa0e085fdfc442e6bdc40303a705bee009c [formerly ca6ea628017592792f06355df43148e15895c16e]
Former-commit-id: 67bd0c711b7b5d4c6a62b11454d5783b445d4401

+++ id/src/main/java/com/simprints/id/services/sync/events/master/models/EventSyncWorkerState.kt
@@ -11,7 +11,7 @@ sealed class EventSyncWorkerState(val state: String) {
     class Failed(
         val failedBecauseCloudIntegration: Boolean = false,
         val failedBecauseBackendMaintenance: Boolean = false,
-        val estimatedOutage: Int = 0
+        val estimatedOutage: Long = 0L
     ) :
         EventSyncWorkerState("Failed")
 
@@ -23,7 +23,7 @@ sealed class EventSyncWorkerState(val state: String) {
             state: WorkInfo.State,
             failedBecauseCloudIntegration: Boolean = false,
             failedBecauseBackendMaintenance: Boolean = false,
-            estimatedOutage: Int = 0
+            estimatedOutage: Long = 0L
         ) =
             when (state) {
                 WorkInfo.State.ENQUEUED -> Enqueued
